requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.5
django-configurations==2.4.1
django-environ==0.11.2
gunicorn==21.2.0
packaging==23.2
psycopg2==2.9.8
sqlparse==0.4.4
typing_extensions==4.8.0
whitenoise==6.5.0
